백엔드 인증 모듈 개선: UserRepository에 커스텀 인터페이스 및 구현체 추가로 QueryDSL을 활용한 유저 정보 업데이트 기능 구현, AuthUserDTO에 성별 및 상태 필드 타입 변경, UserController에서 유저 정보 업데이트 메서드 수정으로 코드 일관성 향상.

+++ back-end/auth/src/main/kotlin/com/placely/auth/service/UserService.kt
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional
 class UserService(
     private val userRepository: UserRepository
 ) {
-    fun updateUserInfo(authUserDTO: AuthUserDTO): Int {
+    fun updateUserInfo(authUserDTO: AuthUserDTO): Long {
         val result = userRepository.updateUserInfo(authUserDTO)
 
         if (result > 0) {
